+++ Assets/DeltaDNA/Triggers/EventAction.cs
@@ -35,6 +35,7 @@ namespace DeltaDNA {
 
         private readonly GameEvent evnt;
         private readonly ReadOnlyCollection<EventTrigger> triggers;
+        private readonly Settings settings;
         private readonly ActionStore store;
 
         private readonly List<EventActionHandler> handlers =
@@ -43,10 +44,12 @@ namespace DeltaDNA {
         internal EventAction(
             GameEvent evnt,
             ReadOnlyCollection<EventTrigger> triggers,
-            ActionStore store) {
+            ActionStore store, 
+            Settings settings) {
 
             this.evnt = evnt;
             this.triggers = triggers;
+            this.settings = settings;
             this.store = store;
         }
 
@@ -66,18 +69,25 @@ namespace DeltaDNA {
         /// Evaluates the registered handlers against the event and triggers
         /// associated for the event.
         /// </summary>
-        public void Run() {
+        public void Run(){
+            bool handledImageMessage = false;
             foreach (var trigger in triggers) {
                 if (trigger.Evaluate(evnt)) {
                     foreach (var handler in handlers) {
-                        if (handler.Handle(trigger, store)) return;
+                        if (handledImageMessage && "imageMessage".Equals(trigger.GetAction())) break;
+                        if (handler.Handle(trigger, store)) {
+                            if (!settings.MultipleActionsForEventTriggerEnabled) return;
+                            if ("imageMessage".Equals(trigger.GetAction())) handledImageMessage = true;
+                            break;
+                        }
+
                     }
                 }
             }
         }
 
         internal static EventAction CreateEmpty(GameEvent evnt) {
-            return new EventAction(evnt, EMPTY_TRIGGERS, null);
+            return new EventAction(evnt, EMPTY_TRIGGERS, null, null);
         }
     }
 }